1.0.0.6版本完成 + 实现Rotation。  + 实现Transform。  + 实现ConvertCoordinates。  + 实现Algebra。  + ThreadingCoreRenderEngine工程整理。  + 实现Arc2。  + ThreadingCoreRenderEngineTesting工程整理。  + ThreadingCoreRenderEngineAssist工程整理。  + ThreadingCoreRenderEngineAssistTesting工程整理。  + 实现Polygon2。  + 实现Sector2。  + 实现Primitives2D。  + 实现Circle3。  + 实现ConvexMesh3。  + 实现AlignedBox。  + 实现ConvexPolyhedron3。  + 实现Line。  + 实现Cylinder3。  + 实现Ellipse3。  + 实现Frustum3。  + 实现Rectangle。  + 实现Lozenge3。  + 实现Polyhedron3。  + 实现SingularValueDecomposition。  + 实现Hyperplane。  + 实现Tetrahedron3。  + 实现Torus3。  + 实现Primitives3D。  + 实现CanonicalBox。  + 实现Segment。  + 实现Capsule。  + 实现Ray。  + 实现UniqueVerticesSimplices。  + 实现Cone。  + 实现Halfspace。  + 实现SymmetricEigensolver。  + 实现Hyperellipsoid。  + 实现Hypersphere。  + 实现OrientedBox。  + 实现Triangle。  + 实现SegmentMesh。  + 实现PrimitivesND。  + 实现Primitives。

+++ Engine/Code/CoreTools/DataTypes/ThreadSafeQueueDetail.h
@@ -11,6 +11,7 @@
 #define CORE_TOOLS_DATA_TYPE_THREAD_SAFE_QUEUE_DETAIL_H
 
 #include "ThreadSafeQueue.h"
+#include "System/Helper/PragmaWarning/NumericCast.h"
 #include "CoreTools/Helper/ClassInvariant/CoreToolsClassInvariantMacro.h"
 #include "CoreTools/Helper/ExceptionMacro.h"
 
@@ -112,7 +113,7 @@ bool CoreTools::ThreadSafeQueue<Element>::Push(ParamType element)
 
     CORE_TOOLS_CLASS_IS_VALID_9;
 
-    if (queue.size() < maxNumElements)
+    if (boost::numeric_cast<int>(queue.size()) < maxNumElements)
     {
         queue.emplace(element);
